Cantilever FRP-Confined Circular Reinforced Concrete Column under Cyclic Lateral Loading.
The cantilever column was modeled by a linear beam element with its stiffness corresponding to flexural yielding and by a fiber element at the plastic hinge which is used in order to capture the flexural hysteretic behavior. The length of the fiber element was assumed to be half of the column’s diameter. A rotational spring at the bottom of the column represents the longitudinal bar pullout from the footing and was assumed to have an elastic stiffness. According to FRPConfinedConcrete model, the averaged response of the two different regions - concrete core (confined by both the FRP & the existing reinforcement) and concrete cover (confined only with the FRP wrap) - in the cross-section allows the assignment of a unique stress-strain law (FRPConfinedConcrete) to all the concrete fibers/layers of the circular section.